// frontend/src/app/api/generated/models/exam-analysis-summary.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { TopicSummaryLine } from '../models/topic-summary-line';
export interface ExamAnalysisSummary {
  headline: string;
  needs_review: Array<TopicSummaryLine>;
  overall_score_pct: number;
  strengths: Array<TopicSummaryLine>;
  weaknesses: Array<TopicSummaryLine>;
}